+++ kernel/src/sys/syscalls/exit.c
@@ -2,7 +2,7 @@
 #include <sched/sched.h>
 
 u64 syscall_exit(syscall_args a) {
-  int status = (int)a.arg1;
+  u64 status = (u64)a.arg1;
   sched_exit(status);
   return 0;
 }
