+++ kernel/src/sys/syscalls/kill.c
@@ -6,7 +6,7 @@ u64 syscall_kill(syscall_args a) {
   task_ctrl* task = sched_get_task((u64)a.arg1);
   if (task == NULL)
     return 1;
-  u8 signal = (u8)a.arg2;
+  u64 signal = (u64)a.arg2;
   sig_raise((signal ? signal : SIGKILL));
   sched_kill(task);
   return 0;
